requirements.txt
selenium
webdriver-manager
requests
beautifulsoup4
python-dotenv
sanitize-filename